requirements.txt
google-generativeai>=0.3.0
SpeechRecognition>=3.10.0
pyttsx3>=2.90
pyaudio>=0.2.11
python-dotenv>=1.0.0
gTTS>=2.3.0
pygame>=2.5.0 